Refactor: components structure
Refactor: move Card, Chip, SnackBar into common dir
Refactor: separate style from component in IssueViewer, Repository, Search
