+++ components/Error.js
@@ -3,7 +3,7 @@
 import React from 'react';
 import type { Node } from 'react';
 
-import SnackBar from './SnackBar';
+import SnackBar from './common/SnackBar';
 
 const Error: () => Node = () => {
   return <SnackBar onPress={() => {}} text="에러가 발생했습니다." />;
