+++ components/common/Chip.js
@@ -8,7 +8,7 @@ import type {
   TextStyleProp,
 } from 'react-native/Libraries/StyleSheet/StyleSheet';
 
-import { COLORS } from '../constants';
+import { COLORS } from '../../constants';
 
 type Props = {
   text: string,
