+++ components/common/SnackBar.js
@@ -8,7 +8,7 @@ import type {
   ViewStyleProp,
 } from 'react-native/Libraries/StyleSheet/StyleSheet';
 
-import { COLORS } from '../constants';
+import { COLORS } from '../../constants';
 
 type Props = {
   onPress: () => void,
